refactor: move CastMemberDtoMapper to actor package
